Move backends into separate modules
This simplifies the code a little (especially in next version when we
drop the hardcoded list). Currently unused and untested backends are
removed.

JIRA: PDC-2585
